move methods from dc_crate into their own modules
